some progress

+++ SegmentationDataset.py
@@ -1,4 +1,5 @@
 from torchvision.datasets import VisionDataset
+import torch
 import os
 from torchvision.io import read_image, ImageReadMode
 from torchvision.transforms.functional import convert_image_dtype
@@ -13,7 +14,7 @@ class SegmentationDataset(VisionDataset):
     def __getitem__(self, idx):
         x = self.samples[idx]  # filename
         y = self.annotations[x]["segmentation"]
-        return convert_image_dtype(read_image(os.path.join(self.root, x), ImageReadMode.GRAY)), self.target_transform(y)
+        return convert_image_dtype(read_image(os.path.join(self.root, x), ImageReadMode.GRAY), dtype=torch.float32), self.target_transform(y)
 
     def __len__(self):
         return len(self.samples)
